+++ Reach/src/utopia/reach/component/label/empty/EmptyLabel.scala
@@ -3,8 +3,10 @@ package utopia.reach.component.label.empty
 import utopia.firmament.context.color.StaticColorContext
 import utopia.firmament.drawing.immutable.{BackgroundDrawer, CustomDrawableFactory}
 import utopia.firmament.drawing.template.CustomDrawer
+import utopia.firmament.drawing.view.BackgroundViewDrawer
 import utopia.firmament.model.stack.StackSize
 import utopia.flow.collection.immutable.Empty
+import utopia.flow.view.template.eventful.Changing
 import utopia.paradigm.color.ColorLevel.Standard
 import utopia.paradigm.color.{Color, ColorLevel, ColorRole}
 import utopia.reach.component.factory.ComponentFactoryFactory.Cff
@@ -13,30 +15,46 @@ import utopia.reach.component.factory.contextual.{ColorContextualFactory, Contex
 import utopia.reach.component.hierarchy.ComponentHierarchy
 import utopia.reach.component.template.{ConcreteCustomDrawReachComponent, PartOfComponentHierarchy}
 
-object EmptyLabel extends Cff[EmptyLabelFactory]
-{
-	override def apply(hierarchy: ComponentHierarchy) = new EmptyLabelFactory(hierarchy)
-}
-
-case class EmptyLabelFactory(hierarchy: ComponentHierarchy)
+case class EmptyLabelFactory(hierarchy: ComponentHierarchy, customDrawers: Seq[CustomDrawer] = Empty)
 	extends FromContextFactory[StaticColorContext, ContextualEmptyLabelFactory] with PartOfComponentHierarchy
+		with CustomDrawableFactory[EmptyLabelFactory]
 {
 	// IMPLEMENTED  --------------------------------
 	
-	override def withContext(context: StaticColorContext) = ContextualEmptyLabelFactory(hierarchy, context)
+	override def withContext(context: StaticColorContext) =
+		ContextualEmptyLabelFactory(hierarchy, context, customDrawers)
+	override def withCustomDrawers(drawers: Seq[CustomDrawer]): EmptyLabelFactory = copy(customDrawers = drawers)
 	
 	
 	// OTHER    ------------------------------------
 	
 	/**
-	 * Creates a new empty label
-	 * @param stackSize Stack size of this label
-	 * @param customDrawers Custom drawers to assign to this label (default = empty)
-	 * @return A new empty label
-	 */
-	def apply(stackSize: StackSize, customDrawers: Seq[CustomDrawer] = Empty) =
-		new EmptyLabel(hierarchy, stackSize, customDrawers)
+	  * Creates a new empty label
+	  * @param stackSize Stack size of this label
+	  * @return A new empty label
+	  */
+	def apply(stackSize: StackSize) = new EmptyLabel(hierarchy, stackSize, customDrawers)
+	/**
+	  * Creates a new empty label
+	  * @param stackSize     Stack size of this label
+	  * @param customDrawers Custom drawers to assign to this label
+	  * @return A new empty label
+	  */
+	@deprecated("Deprecated for removal. Please use .withCustomDrawers(...).apply(StackSize) instead", "v1.6")
+	def apply(stackSize: StackSize, customDrawers: Seq[CustomDrawer]): EmptyLabel =
+		withAdditionalCustomDrawers(customDrawers).apply(stackSize)
 	
+	/**
+	  * @param color Background color to assign
+	  * @return Copy of this factory which draws the specified background color
+	  */
+	def withBackground(color: Color) = withCustomBackgroundDrawer(BackgroundDrawer(color))
+	/**
+	  * @param colorPointer A background color pointer
+	  * @return Copy of this factory which draws a background based on the specified pointer's value
+	  */
+	def withBackground(colorPointer: Changing[Color]) =
+		withCustomBackgroundDrawer(BackgroundViewDrawer(colorPointer))
 	/**
 	 * Creates a new empty label with a static background color
 	 * @param color Background color for this label
@@ -44,6 +62,7 @@ case class EmptyLabelFactory(hierarchy: ComponentHierarchy)
 	 * @param customDrawers Additional custom drawers to assign (default = empty)
 	 * @return A new empty label
 	 */
+	@deprecated("Deprecated for removal. Please use .withBackground(Color).apply(StackSize) instead", "v1.6")
 	def withBackground(color: Color, stackSize: StackSize, customDrawers: Seq[CustomDrawer] = Empty) =
 		apply(stackSize, BackgroundDrawer(color) +: customDrawers)
 }
@@ -60,7 +79,7 @@ case class ContextualEmptyLabelFactory(hierarchy: ComponentHierarchy, context: S
 	 * @return A copy of this factory without context information
 	 */
 	@deprecated("Deprecated for removal", "v1.1")
-	def withoutContext = EmptyLabel(hierarchy)
+	def withoutContext = EmptyLabelFactory(hierarchy, customDrawers)
 	
 	
 	// IMPLEMENTED  --------------------------------
@@ -94,6 +113,10 @@ case class ContextualEmptyLabelFactory(hierarchy: ComponentHierarchy, context: S
 		withoutContext.withBackground(context.color.preferring(preferredShade)(role), stackSize, customDrawers)
 }
 
+object EmptyLabel extends Cff[EmptyLabelFactory]
+{
+	override def apply(hierarchy: ComponentHierarchy) = EmptyLabelFactory(hierarchy)
+}
 /**
  * A simple immutable empty label
  * @author Mikko Hilpinen